move tests from mutations to utils
